requirements.txt
disnake
coloredlogs
aiohttp
asyncpg
importlib
PyYAML
python-dotenv
